requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
langchain==0.1.20
langchain-openai==0.0.8
openai>=1.6.0
aiofiles
requests
python-multipart
websockets